==> 01_tratar_dados.ipynb <==
# 1. Importar Bibliotecas

#Importar bibliotecas
import pandas as pd
import numpy as np

# 2. Carregar Dados e Tratamento Inicial

#Carregar dados
base_usuarios = pd.read_csv('base_usuarios_final.csv')
base_transacoes = pd.read_csv('base_transacoes_final.csv')

## 2.1. Base Usuários

#Checar nulos
100 * base_usuarios.isnull().sum() / len(base_usuarios)

Dado o tamanho da base e o pequeno percentual de registros com valores nulos, vamos adotar a simples ação de remover os registros em questão.

#Remover registros com campos nulo
base_usuarios.dropna(inplace = True)

#Verificar tipos dos campos
base_usuarios.dtypes

#Tratar campo de data / remover informações de horário
base_usuarios['createdAt'] = [np.datetime64(data[:10]) for data in base_usuarios['createdAt']]

#Renomear colunas
novas_colunas = {
    'user_id': 'id_usuario', 
    'createdAt': 'data_criacao_conta',
    'state_anom': 'estado_anonimo',
    'is_state_capital': 'flag_capital_do_estado',
    'age_range': 'faixa_idade',
    'gender': 'genero'}

base_usuarios.rename(columns = novas_colunas, inplace = True)

## 2.2. Base Transações

#Checar nulos
100 * base_transacoes.isnull().sum() / len(base_usuarios)

len(base_transacoes)

base_transacoes.isnull().sum()

#Remover registros com campos nulo
base_transacoes.dropna(inplace = True)

#Verificar tipos dos campos
base_transacoes.dtypes

#Tratar campo de data
base_transacoes['createdat'] = [np.datetime64(data) for data in base_transacoes['createdat']]

#Renomear colunas
novas_colunas = {
    'createdat': 'data_transacao',
    'type': 'grupo_transacao',
    'transactiontype_anom': 'tipo_transacao_anonimo',
    'amount_anom': 'valor',
    'user_id': 'id_usuario'}

base_transacoes.rename(columns = novas_colunas, inplace = True)

# 3. Tratamento dos Dados

## 3.1. Consolidação dos Dados de Transações

A base de transações contém todas as 2 primeiras transações dos usuários. O primeiro passo na nossa análise é consolidar os dois registros em um só, de forma a ter um entendimento melhor do comportamento inicial do usuário.

#Criar campo de ordenação das transações
base_transacoes['ordem_transacao'] = base_transacoes.groupby(by = 'id_usuario').cumcount() + 1

#Consolidar dados
base_transacoes = base_transacoes[base_transacoes['ordem_transacao'] == 1].join(
    on = 'id_usuario',
    other = base_transacoes[base_transacoes['ordem_transacao'] == 2].set_index('id_usuario'),
    rsuffix = '_2')

#Renomear colunas relativas à primeira transação
novas_colunas = {
    'data_transacao': 'data_transacao_1',
    'grupo_transacao': 'grupo_transacao_1',
    'tipo_transacao_anonimo': 'tipo_transacao_anonimo_1',
    'valor': 'valor_1'}

base_transacoes.rename(columns = novas_colunas, inplace = True)

len(base_transacoes)

#Remover campos ordem_transacao
base_transacoes.drop(columns = ['ordem_transacao', 'ordem_transacao_2'], inplace = True)

#Criar campo calculado dias_entre_transacoes
for col in ['data_transacao_1', 'data_transacao_2']:
    base_transacoes[col] = [np.datetime64(str(data)[:10]) for data in base_transacoes[col]]
    
base_transacoes['dias_entre_transacoes'] = [(tx_2 - tx_1) / np.timedelta64(1, 'D') for tx_2, tx_1 in 
                                            zip(base_transacoes['data_transacao_2'], base_transacoes['data_transacao_1'])]

#Criar compo saldo_restante
saldo_1 = [valor if grupo == 'cashin' else - valor for valor, grupo 
           in zip(base_transacoes['valor_1'], base_transacoes['grupo_transacao_1'])]
saldo_2 = [valor if grupo == 'cashin' else - valor for valor, grupo 
           in zip(base_transacoes['valor_2'], base_transacoes['grupo_transacao_2'])]

base_transacoes['saldo_restante'] = [val_1 + val_2 + 5 for val_1, val_2 in zip(saldo_1, saldo_2)]

print('Percentual de registros com saldo restante negativo: {:.2f}%'.format(
    100 * len(base_transacoes[base_transacoes['saldo_restante'] < 0]) / len(base_transacoes)))

Na criação do campo saldo_restante, somamos R$ 5 ao saldo final de todos os usuários por conta de um bônus que é aplicado aos usuários por conta de sua primeira transação.

Adicionalmente, cerca de 20% dos registros apresentam o saldo em conta restante negativo, mesmo considerando o acréscimo do bônus. Isso decorre do fato que, no processo de anonimização dos dados, foi incluído certo nível de ruído nos valores dos campos. Com isso, não faz sentido realizar essa operação com os dois atributos, dado que os efeitos do ruído podem se cancelar ou somar. Ficamos então com a decisão de remover o campo da nossa análise.

#Remover campo saldo_restante
base_transacoes.drop(columns = 'saldo_restante', inplace = True)

## 3.2. Cruzar Dados de Usuários e Transações

#Cruzar dados das bases
base_atividade = base_transacoes.join(on = 'id_usuario', other = base_usuarios.set_index('id_usuario')) 

#Criar campo calculado dias_ate_primeira_transacao
base_atividade['dias_ate_primeira_transacao'] = [(tx - conta) / np.timedelta64(1, 'D') for tx, conta in 
                                                 zip(base_atividade['data_transacao_1'], base_atividade['data_criacao_conta'])]

#Reordenar colunas
colunas = list(base_atividade.columns)
colunas.remove('id_usuario')
colunas = ['id_usuario'] + colunas

base_atividade = base_atividade[colunas] 

base_atividade.head()

#Consolidar dados com target
base_target = pd.read_csv('base_target_final.csv')
base_target['flag_retido'] = [1 if mes >=4 else 0 for mes in base_target['meses_ativo']]

base_atividade = base_atividade.join(on = 'id_usuario', other = base_target.set_index('user_id')['flag_retido'])

## 3.3. Tratamento Adicional dos Dados

Com o cruzamento das duas fontes de dados, é possível que algum tratamento final seja necessário nos dados. Vamos agora verificar essa possibilidade.

#Verificar existência de registros nulos na base consolidada
nulos = base_atividade.isnull().sum()
100 * nulos / len(base_atividade)

len(base_atividade)

nulos

O que percebemos aqui são uma série de campos com registros nulos. No entanto, olhando com mais atenção, podemos identificar que os campos com valores faltando derivam de duas fontes:

* Dados da segunda transação dos usuários;
* Dados cadastrais.

No primeiro caso, a escolha é óbvia. Ainda que representem quase 10% de todos os registros da base, precisamos remover todos as linhas com valores nulos relacionados à segunda transação. Como queremos medir a probabilidade de retenção dos usuários, não há dúvida de que usuários que não chegaram sequer a fazer 2 operações não podem ser considerados retidos.

No segundo caso, a opção mais simples é também a remoção, especialmente considerando a baixa representatividade no tamanho total da base de dados. Analisando o caso mais à fundo, relembramos que, na coleta dos dados cadastrais dos usuários, removemos todos aqueles clientes em situação irregular. De maneira mais clara, foram removidos todos os usuários bloqueados por suspeita/confirmação de fraude. A inexistência de dados para esses usuários é um indicativo de que estariam dentro dessa classificação. Assim, fica mais embasada a escolha pela deleção dos registros.

#Remover registros com dados nulos
base_atividade.dropna(inplace = True)

len(base_atividade)

#Salvar base de dados tratada
base_atividade.to_csv('base_final.csv', index = False)
